pyproject.toml
[project]
name = "mummymaze-rl"
version = "0.1.0"
description = "PopCap's Mummy Maze Deluxe as a JAX RL testbed"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beartype>=0.22.9",
    "diffrax>=0.6.0",
    "equinox>=0.13.2",
    "jax[cuda13]>=0.8.2",
    "jaxtyping>=0.3.5",
    "joblib>=1.5.3",
    "matplotlib>=3.10.8",
    "mummy-maze-parser",
    "optax>=0.2.6",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.13",
    "ty>=0.0.12",
    "vulture>=2.14",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.ruff]
indent-width = 2

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # pyflakes
  "ANN", # flake8-annotations
]
ignore = [
  "E731",   # lambdas are fine for FP-style composition
  "F722",   # jaxtyping dimension strings look like syntax errors
  "F821",   # jaxtyping dimension names look like undefined variables
  "ANN401", # Any is needed for dtype parameters (no good JAX dtype type)
]

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = false
ignore-fully-untyped = false
suppress-none-returning = true

[tool.ruff.format]
indent-style = "space"

[tool.ty.rules]
# Upgrade warnings to errors for strict mode
ambiguous-protocol-member = "error"
deprecated = "error"
ignore-comment-unknown-rule = "error"
invalid-ignore-comment = "error"
possibly-missing-attribute = "error"
possibly-missing-implicit-call = "error"
redundant-cast = "error"
# Enable rules that are off by default
division-by-zero = "error"
possibly-unresolved-reference = "error"

[tool.vulture]
paths = ["src"]
min_confidence = 60

[tool.uv.sources]
mummy-maze-parser = { git = "https://github.com/jonahshader/mummy-maze-parser" }
